new_tube 0.3
The bot is no longer controllable with telegram. It will only write, but not read any message.
The bot will now always use the default channel to connect to. It is designed to be for personal use only.
new_tube now has a config. The minutes until the next fetch can be configured here.